requirements.txt
asgiref==3.5.2
boto3==1.24.56
botocore==1.27.56
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==4.0.5
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-storages==1.13.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
Pillow==9.2.0
psycopg2==2.9.3
python-dateutil==2.8.2
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
